% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging_table.R
\name{aged_analysis}
\alias{aged_analysis}
\title{Calculates the number of outstanding days of invoices and assigns them to
chosen categories of overdue days}
\usage{
aged_analysis(
  df,
  due_date,
  report_date = Sys.Date(),
  categories = c(0, 30, 60, 90)
)
}
\arguments{
\item{df}{\emph{a data frame} including columns indicating the following:
debtors ID/name, invoice number, invoice amount, and invoice due date.
Usually this is going to be a Debtors Open Item Report, including all
invoices that have not been (fully) paid to a specific date.}

\item{due_date}{\emph{a column name} in the data frame indicating the due
date. Column name should be assigned to the parameter in quotation marks,
e.g. "Due_Date".}

\item{report_date}{\emph{a report date} which is the benchmark if an invoice
is overdue or not, usually the same date that the Debtors Open Item Report
has been generated from. The default is the system date. The format must be
"YYYY-MM-DD".}

\item{categories}{\emph{a vector of numbers} listing the days to be used as
overdue bins. By default it is 0 days, 30 days, 60 days and 90 days. Format
must be numeric values only, e.g. c(0, 30, 60, 90)}
}
\value{
a dataframe which has two additional columns: \describe{
\item{\strong{days_overdue}}{indicating the number of days between
report_date and due_date} \item{\strong{category}}{indicating the bin the
days_overdue have been sorted in based on user's categories vector} }

@examples
aged_analysis(dataframe, "Due_Date", "2022-12-31")
aged_analysis(dataframe, "Due_Date", categories=c(0,60,120))
}
\description{
\code{aged_analysis} is the first part of a two-step workflow to create a
Accounts Receivable Aged Analysis, widely used from accountants, auditors and
management to monitor accounts receivables. It categorizes the invoices
based on how long invoices have been outstanding.
}
\details{
First the user should use \code{aged_analysis} to calculate the days
outstanding and categorize them to the chosen bins of days overdue. By
default, the following categories will be used: \enumerate{ \item Not overdue
yet ("-Inf_to_0") \item Overdue between 0 and 30 days ("0_to_30") \item
Overdue between 30 and 60 days ("30_to_60") \item Overdue between 60 and 90
days ("60_to_90") \item Overdue more than 90 days ("90_to_Inf").}

It is recommended to use \code{\link{aging_report}} as the next step.
}
