% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging_table.R
\name{aging_report}
\alias{aging_report}
\title{Creates a Accounts Receivable Aged Analysis Report}
\usage{
aging_report(
  df,
  open_amount,
  customer,
  invoice_number,
  include.credit.notes = TRUE
)
}
\arguments{
\item{df}{\emph{a data frame} including columns indicating the following:
debtors ID/name, invoice number, invoice amount, and invoice due date.
Usually this is going to be a Debtors Open Item Report, including all
invoices that have not been (fully) paid to a specific date.}

\item{open_amount}{\emph{a column name} in the data frame indicating the open
amount of the invoice. Column name should be assigned to the parameter in
quotation marks, e.g. "Net_Amount". It is recommended to check if the open
amounts are all in the same currency.}

\item{customer}{\emph{a column name} in the data frame indicating the ID or
name of the customer that the invoice belongs to. Column name should be
assigned to the parameter in quotation marks, e.g. "Customer". Customer
names or IDs should be unique, so if there are multiple debtors with the
same name, it might be useful to use the customer ID instead.}

\item{invoice_number}{\emph{a column name} in the data frame indicating the
invoice number. Column name should be assigned to the parameter in
quotation marks, e.g. "Invoice_Number".}

\item{include.credit.notes}{\emph{a logical}, by default set to TRUE. This
logical indicates if amounts that are negative (usually representing credit
notes) should be excluded or not. It might be useful to exclude credit
notes if the aged report is used to calculate the bad debt expenses. It
might be useful to include credit notes if the aged report is used to
monitor on a customer level how much of the balance is overdue. This
depends on the usage of the report and should be evaluated on an individual
basis of the user.}
}
\value{
a data frame which has the following contents: \describe{
\item{\strong{Customer names.}}{Each customer only appears once in the
list.} \item{\strong{A column for each category chosen in
\code{\link{aged_analysis}}}}{Indicating the amount which is due for a
specific customer in this duration.}
\item{\strong{Total}}{Indicating the total amount which is due for this customer.}}
}
\description{
\code{aging_report} is the second part of a two-step workflow to created a
Accounts Receivable Aged Analysis, widely used from accountants, auditors
and management to monitor accounts receivables. It categorizes the invoices
based on how long invoices have been outstanding. For the first step, please
refer to \code{\link{aged_analysis}}.
}
