% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_home_currency.R
\name{FX_rate_convert}
\alias{FX_rate_convert}
\title{Converts foreign currencies in a data frame to a chosen currency by the user}
\usage{
FX_rate_convert(
  data,
  FC_column,
  amount,
  new_currency,
  report_date = as.Date(Sys.Date())
)
}
\arguments{
\item{data}{\emph{a data frame} including columns indicating the following:
foreign exchange course identifier and the amounts to be translated into the new currency.}

\item{FC_column}{\emph{a column name} in the data frame indicating the foreign exchange course identifier. See \code{priceR::currencies()} for supported currency codes.}

\item{amount}{\emph{a column name} in the data frame indicating the amounts which should be translated into another currency.}

\item{new_currency}{\emph{a currency code} indicating which currency the amounts should be translate into. See \code{priceR::currencies()} for supported currency codes.}

\item{report_date}{\emph{a date} to which the amounts should be translated into another currency. The default is the system date. The format must be
"YYYY-MM-DD".}
}
\value{
a dataframe which has two additional columns: \describe{
\item{\strong{FX_rate}}{the applied exchange rate} \item{\strong{translated_amount}}{the amount translated to the chosen currency} }

@examples
currencies <- data.frame(
currency = c("USD", "DKK", "USD", NA),
price = c(100, 329, 23, 799),
date = as.Date(c("2020-01-01", "2020-02-01", "2020-03-01", "2020-03-01")))

FX_rate_convert(data = currencies, amount = "price", FC_column = "currency", new_currency = "EUR")
FX_rate_convert(data = currencies, amount = "price", FC_column = "currency", new_currency = "USD", report_date = "2022-06-30")
}
\description{
\code{FX_rate_convert} downloads foreign exchange courses from the European Central Bank's API to convert foreign currencies to a chosen currency by the user and documents the exchange rate used.
}
